+++ react/src/lib/SpaceListMeeting/index.js
@@ -23,7 +23,9 @@ class SpaceListMeeting extends React.PureComponent {
     const {
       attendees,
       buttonAriaLabel,
+      buttonClassName,
       buttonLabel,
+      buttonTabIndex,
       childrenLeft,
       childrenRight,
       className,
@@ -119,8 +121,10 @@ class SpaceListMeeting extends React.PureComponent {
             color={isMessagingOnlyShare ? 'blue' : 'green'}
             ariaLabel={buttonAriaLabel ? buttonAriaLabel : buttonLabel}
             children={buttonLabel}
+            className={buttonClassName}
             onClick={this.handleButtonClick}
             size={28}
+            tabIndex={buttonTabIndex}
           />
         )}
       </ListItemSection>,
@@ -157,10 +161,14 @@ SpaceListMeeting.propTypes = {
   ),
   /** @prop Aria label string for button | buttonLabel */
   buttonAriaLabel: PropTypes.string,
+  /** @prop Optionsl HTML Class string for button | '' */
+  buttonClassName: PropTypes.string,
   /** @prop Label string for button | '' */
   buttonLabel: PropTypes.string,
   /** @prop Callback function invoked when user clicks on button | null */
   buttonOnClick: PropTypes.func,
+  /** @prop Tab index for button | '0' */
+  buttonTabIndex: PropTypes.string,
   /** @prop Children nodes to render for left section | null */
   childrenLeft: PropTypes.node,
   /** @prop Children nodes to render for right section | null */
@@ -190,8 +198,10 @@ SpaceListMeeting.propTypes = {
 SpaceListMeeting.defaultProps = {
   attendees: [],
   buttonAriaLabel: null,
+  buttonClassName: '',
   buttonLabel: '',
   buttonOnClick: null,
+  buttonTabIndex: '0',
   childrenLeft: null,
   childrenRight: null,
   className: '',