+++ web-components/.storybook/main.ts
@@ -16,7 +16,9 @@ const mergeUnique = (merger: Array<any>, mergee?: Array<any>) =>
 module.exports = {
   stories: [
     "../src/components/**/*.stories.ts",
-    "../src/components/**/*.stories.mdx"
+    "../src/components/**/*.stories.mdx",
+    "../src/internal-components/color-table/ColorTable.stories.ts",
+    "../src/internal-components/color-table/*.stories.mdx"
   ],
 
   addons: [