+++ web-components/src/components/checkbox/Checkbox.stories.ts
@@ -14,7 +14,7 @@ import { boolean, text, withKnobs } from "@storybook/addon-knobs";
 import { html } from "lit-element";
 
 export default {
-  title: "Checkbox",
+  title: "Components/Checkbox",
   component: "md-checkbox",
   decorators: [withKnobs, withA11y],
   parameters: {