+++ core/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [19.9.45](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/core@19.9.44...@momentum-ui/core@19.9.45) (2021-03-19)
+
+**Note:** Version bump only for package @momentum-ui/core
+
+
+
+
+
+## [19.9.44](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/core@19.9.43...@momentum-ui/core@19.9.44) (2021-03-17)
+
+**Note:** Version bump only for package @momentum-ui/core
+
+
+
+
+
 ## [19.9.43](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/core@19.9.42...@momentum-ui/core@19.9.43) (2021-03-04)
 
 **Note:** Version bump only for package @momentum-ui/core