+++ react/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [23.22.1](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/react@23.22.0...@momentum-ui/react@23.22.1) (2021-03-19)
+
+**Note:** Version bump only for package @momentum-ui/react
+
+
+
+
+
+# [23.22.0](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/react@23.21.9...@momentum-ui/react@23.22.0) (2021-03-17)
+
+
+### Features
+
+* **spacelistmeeting:** add props for button tabIndex and className ([#787](https://github.com/momentum-design/momentum-ui/issues/787)) ([6c22595](https://github.com/momentum-design/momentum-ui/commit/6c225952468c615d0f0935bb92906d240bcd12b4))
+
+
+
+
+
 ## [23.21.9](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/react@23.21.8...@momentum-ui/react@23.21.9) (2021-03-04)
 
 **Note:** Version bump only for package @momentum-ui/react